修复地物分类lut无法保存的bug

+++ backend/applications/interface/semantic_segmentation.py
@@ -20,6 +20,6 @@ def execute(model_path, data_path, out_dir, test_names):
     for idx, im in zip(range(len(image_list)), ims):
         im = lut[im]
         new_name = md5_name(test_names[idx])
-        imsave(osp.join(out_dir, new_name), im)
+        imsave(osp.join(out_dir, new_name), np.uint8(im))
         temps.append(generate_url + new_name)
     return temps
